backend/requirements.txt
fastapi 
uvicorn 
pydantic-ai[logfire]
logfire[asyncpg]
google-cloud-bigquery 
supabase 
openai
pydantic-settings  # Config
SQLAlchemy
databases[asyncpg]
python-jose # JWT
passlib
bcrypt